+++ include/unity/scopes/OnlineAccountClient.h
@@ -37,6 +37,11 @@ namespace internal
 class OnlineAccountClientImpl;
 }
 
+namespace testing
+{
+class OnlineAccountClientTest;
+}
+
 /**
 \brief A simple interface for integrating online accounts access and monitoring into scopes.
 
@@ -68,16 +73,13 @@ public:
     /**
     \brief Indicates whether an external main loop already exists, or one should be created internally.
 
-    A running main loop is essential in order to receive service update callbacks. When in doubt, set
-    RunInExternalMainLoop and just use refresh_service_statuses() and get_service_statuses() to obtain
-    service statuses.
+    A running main loop is essential in order to receive service updates from the online accounts
+    backend. When in doubt, set to CreateInternalMainLoop.
     */
     enum MainLoopSelect
     {
-        RunInExternalMainLoop,  ///< An external main loop already exists or the service update
-                                ///  callback is not required.
-        CreateInternalMainLoop  ///< An external main loop does not exist and the service update
-                                ///  callback is required.
+        RunInExternalMainLoop,    ///< An external main loop already exists and is running.
+        CreateInternalMainLoop,   ///< An external main loop does not exist.
     };
 
     /**
@@ -91,7 +93,7 @@ public:
     OnlineAccountClient(std::string const& service_name,
                         std::string const& service_type,
                         std::string const& provider_name,
-                        MainLoopSelect main_loop_select = RunInExternalMainLoop);
+                        MainLoopSelect main_loop_select = CreateInternalMainLoop);
 
     /// @cond
     ~OnlineAccountClient();
@@ -160,6 +162,8 @@ public:
 
 private:
     std::unique_ptr<internal::OnlineAccountClientImpl> p;
+
+    friend class testing::OnlineAccountClientTest;
 };
 
 } // namespace scopes